% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_cjpg_covid}
\alias{da_cjpg_covid}
\title{da_cjpg_covid}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30766 rows and 10 columns.
}
\usage{
da_cjpg_covid
}
\description{
da_cjpg_covid
}
\keyword{datasets}
