% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_cpopg_covid}
\alias{da_cpopg_covid}
\title{da_cpopg_covid}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 142351 rows and 19 columns.
}
\usage{
da_cpopg_covid
}
\description{
da_cpopg_covid
}
\keyword{datasets}
